requirements.txt
llama-cpp-python
chromadb==0.4.22
langchain==0.1.1
pydantic==2.5.3
transformers==4.36.2
sentence-transformers==2.2.2
pypdf2==3.0.1
ebooklib==0.18
sentencepiece==0.1.99
setuptools==67.7.2
loguru
python-dotenv
accelerate==0.22.0
protobuf==3.20.2
termcolor
openai==1.8.0
einops # required for Mosaic models
click 
bitsandbytes==0.42.0
# auto-gptq==0.2.0
InstructorEmbedding==1.0.1
unstructured==0.10.11
pymupdf==1.22.5
streamlit==1.28.0
streamlit-chat==0.1.1
python-docx==0.8.11
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy==1.4.48 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.10" and python_version < "4.0"
sympy==1.11.1 ; python_version >= "3.10" and python_version < "4.0"
tenacity==8.2.3 ; python_version >= "3.10" and python_version < "4.0"
threadpoolctl==3.1.0 ; python_version >= "3.10" and python_version < "4.0"
tiktoken==0.3.3 ; python_version >= "3.10" and python_version < "4.0"
tokenizers==0.15.0; python_version >= "3.10" and python_version < "4.0"
tqdm==4.65.0 ; python_version >= "3.10" and python_version < "4.0"
# transformers==4.29.2 ; python_version >= "3.10" and python_version < "4.0"